add javadocs and more code cleanup
