+++ custos-external-applications/custos-ssl-certificate-manager/src/main/java/org/apache/custos/ssl/certificate/manager/clients/NginxClient.java
@@ -20,7 +20,7 @@
 package org.apache.custos.ssl.certificate.manager.clients;
 
 import org.apache.custos.ssl.certificate.manager.configurations.NginxConfiguration;
-import org.apache.custos.ssl.certificate.manager.helpers.QueryString;
+import org.apache.custos.ssl.certificate.manager.helpers.QueryStringHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -28,22 +28,42 @@ import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.net.URL;
 
+
+/**
+ * Nginx client class to handle nginx related requests
+ */
 public class NginxClient {
     private static final Logger logger = LoggerFactory.getLogger(NginxClient.class);
     private final String acmeChallengeUrl;
 
     public NginxClient(NginxConfiguration config) {
-        acmeChallengeUrl = config.getUrl() + config.getFolderPath();
+        acmeChallengeUrl = config.getUrl() + "/acme/challenge";
     }
 
+
+    /**
+     * This method creates the necessary file in nginx which is needed
+     * to validate the acme challenge
+     *
+     * @param fileName Name of the file
+     * @param fileContent Content of the file
+     * @return true if request is success else false
+     */
     public boolean createAcmeChallenge(String fileName, String fileContent) {
-        QueryString query = new QueryString("file", fileName);
+        QueryStringHelper query = new QueryStringHelper("file", fileName);
         query.add("content", fileContent);
         return send("POST", query.toString());
     }
 
+    /**
+     * This method removes the challenge file created in nginx after
+     * validation process is completed
+     *
+     * @param fileName Name of the file
+     * @return true if request is success else false
+     */
     public boolean deleteAcmeChallenge(String fileName) {
-        QueryString query = new QueryString("file", fileName);
+        QueryStringHelper query = new QueryStringHelper("file", fileName);
         return send("DELETE", query.toString());
     }
