+++ custos-external-applications/custos-ssl-certificate-manager/src/main/resources/application.properties
@@ -18,11 +18,10 @@
 #
 
 NGINX_URL=http://localhost:31515
-NGINX_CHALLENGE_FOLDER_PATH=/acme/challenge
 ACME_URL=acme://pebble/localhost:14000
 ACME_DOMAINS=localhost
 CUSTOS_URL=test
-CUSTOS_PORT=31499
+CUSTOS_PORT=test
 CUSTOS_CLIENT_ID=test
 CUSTOS_CLIENT_SECRET=test
 CUSTOS_OWNER_ID=test
